Add files via upload
Bubble sort, sometimes referred to as sinking sort, is a simple sorting algorithm that repeatedly steps through the list, compares adjacent elements and swaps them if they are in the wrong order. The pass through the list is repeated until the list is sorted.
Bubble sort has a worst-case and average complexity of О(n2), where n is the number of items being sorted.