Add files via upload
Insertion sort is a simple sorting algorithm.The array is virtually split into a sorted and an unsorted part. Values from the unsorted part are picked and placed at the correct position in the sorted part.
Time complexity is O(n*2) where n is the number of elements to be sorted.