Add files via upload
Merge Sort is a Divide and Conquer algorithm. It divides the input array into two halves, calls itself for the two halves, and then merges the two sorted halves.
Time complexity of Merge Sort is  θ(nLogn) in all 3 cases (worst, average and best)  where n is the size of the array.