Add files via upload
Heap sort is a comparison based sorting technique based on Binary Heap data structure. It is similar to selection sort where we first find the maximum element and place the maximum element at the end. We repeat the same process for the remaining elements.
Time complexity of Heap Sort is O(nLogn) where n is the number of elements to be sorted.